Merged PR 12987: Add global css for variables
The components that I started in this branch are still in progress but I'm opening this PR to share the global css setup.

This is more just to get it working. Don't take the variables or naming conventions as best practice. We can probably talk more and come to an agreement on that later.